requirements.txt
httpx>=0.27
fastapi>=0.115
uvicorn>=0.32
websockets>=13.0
textual>=0.89
pyyaml>=6.0
oracledb>=2.4
rich>=13.9
